commit CustomerList
