+++ src/customers/components/CustomerItem.js
